refactor: extract handwriting tool into reusable component
