+++ src/components/handwritingTool/HandwritingTool.vue
