+++ src/components/handwritingTool/HierarchicalMetadataEditor.vue
