+++ src/components/handwritingTool/MetadataEditor.vue
