+++ src/components/handwritingTool/VisualizationToggles.vue
