+++ src/components/handwritingTool/ancillaryVisualizations.ts
