+++ src/components/handwritingTool/core/CanvasItem.ts
