+++ src/components/handwritingTool/core/commandStack.ts
