+++ src/components/handwritingTool/core/commands.ts
