+++ src/components/handwritingTool/core/selectTool.ts
