+++ src/components/handwritingTool/core/selectionStore.ts
