+++ src/components/handwritingTool/core/transformerManager.ts
