+++ src/components/handwritingTool/docs/freehandTool_orig_pseudocode.md
