+++ src/components/handwritingTool/freehandTool.ts
