+++ src/components/handwritingTool/gpuStrokes/constants.ts
