+++ src/components/handwritingTool/gpuStrokes/coordinateUtils.ts
