+++ src/components/handwritingTool/gpuStrokes/debug_thoughts.md
