+++ src/components/handwritingTool/gpuStrokes/drawLifecycleManager.ts
