+++ src/components/handwritingTool/gpuStrokes/drawingScene.ts
