+++ src/components/handwritingTool/gpuStrokes/strokeAnimation.wgsl
