+++ src/components/handwritingTool/gpuStrokes/strokeDataGenerator.ts
