+++ src/components/handwritingTool/gpuStrokes/strokeInterpolator.ts
