+++ src/components/handwritingTool/gpuStrokes/strokeTextureManager.ts
