+++ src/components/handwritingTool/gpuStrokes/strokeTypes.ts
