+++ src/components/handwritingTool/index.ts
@@ -1,0 +1,2 @@
+export { default as HandwritingTool } from './HandwritingTool.vue'
+export * from './appState'
