+++ src/components/handwritingTool/polygonTool.ts
