+++ src/components/handwritingTool/typographyGuides.ts
