+++ src/components/handwritingTool/utils/canvasUtils.ts
